mv function to Util
